restructure mock backends and fix packaging of backend data
