+++ MANIFEST.in
@@ -3,3 +3,4 @@ include qiskit/schemas/*json
 include qiskit/VERSION.txt
 include qiskit/transpiler/passes/mapping/cython/stochastic_swap/*pyx
 include qiskit/transpiler/passes/mapping/cython/stochastic_swap/*pxd
+recursive-include qiskit/test/mock/backends *json
