+++ qiskit/test/mock/__init__.py
@@ -27,22 +27,8 @@ from .fake_backend import FakeBackend
 from .fake_job import FakeJob
 from .fake_qobj import FakeQobj
 
+from .backends import *
+
 from .fake_qasm_simulator import FakeQasmSimulator
-
-from .fake_yorktown import FakeYorktown
-from .fake_tenerife import FakeTenerife
-from .fake_ourense import FakeOurense
-from .fake_vigo import FakeVigo
-from .fake_burlington import FakeBurlington
-from .fake_melbourne import FakeMelbourne
-from .fake_rueschlikon import FakeRueschlikon
-from .fake_tokyo import FakeTokyo
-from .fake_poughkeepsie import FakePoughkeepsie
-from .fake_almaden import FakeAlmaden
-from .fake_singapore import FakeSingapore
-from .fake_johannesburg import FakeJohannesburg
-from .fake_boeblingen import FakeBoeblingen
-
-
 from .fake_openpulse_2q import FakeOpenPulse2Q
 from .fake_openpulse_3q import FakeOpenPulse3Q
