+++ qiskit/test/mock/backends/__init__.py
@@ -1,0 +1,32 @@
+# -*- coding: utf-8 -*-
+
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2019.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+
+"""
+Mocked versions of real quantum backends.
+"""
+
+from .yorktown import FakeYorktown
+from .tenerife import FakeTenerife
+from .ourense import FakeOurense
+from .vigo import FakeVigo
+from .burlington import FakeBurlington
+from .melbourne import FakeMelbourne
+from .rueschlikon import FakeRueschlikon
+from .tokyo import FakeTokyo
+from .poughkeepsie import FakePoughkeepsie
+from .almaden import FakeAlmaden
+from .singapore import FakeSingapore
+from .johannesburg import FakeJohannesburg
+from .boeblingen import FakeBoeblingen
