+++ qiskit/test/mock/backends/almaden/props_almaden.json
