+++ qiskit/test/mock/backends/boeblingen/fake_boeblingen.py
@@ -13,7 +13,7 @@
 # that they have been altered from the originals.
 
 """
-Fake Johannesburg device (20 qubit).
+Fake Boeblingen device (20 qubit).
 """
 
 import os
@@ -21,31 +21,33 @@ import json
 
 from qiskit.providers.models import (GateConfig, QasmBackendConfiguration,
                                      BackendProperties)
-from .fake_backend import FakeBackend
+from qiskit.test.mock.fake_backend import FakeBackend
 
 
-class FakeJohannesburg(FakeBackend):
-    """A fake Johannesburg backend."""
+class FakeBoeblingen(FakeBackend):
+    """A fake Boeblingen backend."""
 
     def __init__(self):
         """
           00 ↔ 01 ↔ 02 ↔ 03 ↔ 04
-           ↕                   ↕
+                ↕         ↕
           05 ↔ 06 ↔ 07 ↔ 08 ↔ 09
            ↕         ↕         ↕
           10 ↔ 11 ↔ 12 ↔ 13 ↔ 14
-           ↕                   ↕
+                ↕         ↕
           15 ↔ 16 ↔ 17 ↔ 18 ↔ 19
         """
-        cmap = [[0, 1], [0, 5], [1, 0], [1, 2], [2, 1], [2, 3], [3, 2], [3, 4], [4, 3], [4, 9],
-                [5, 0], [5, 6], [5, 10], [6, 5], [6, 7], [7, 6], [7, 8], [7, 12], [8, 7], [8, 9],
-                [9, 4], [9, 8], [9, 14], [10, 5], [10, 11], [10, 15], [11, 10], [11, 12], [12, 7],
-                [12, 11], [12, 13], [13, 12], [13, 14], [14, 9], [14, 13], [14, 19], [15, 10],
-                [15, 16], [16, 15], [16, 17], [17, 16], [17, 18], [18, 17], [18, 19], [19, 14],
+        cmap = [[0, 1], [1, 0], [1, 2], [1, 6], [2, 1], [2, 3], [3, 2],
+                [3, 4], [3, 8], [4, 3], [5, 6], [5, 10], [6, 1], [6, 5],
+                [6, 7], [7, 6], [7, 8], [7, 12], [8, 3], [8, 7], [8, 9],
+                [9, 8], [9, 14], [10, 5], [10, 11], [11, 10], [11, 12],
+                [11, 16], [12, 7], [12, 11], [12, 13], [13, 12], [13, 14],
+                [13, 18], [14, 9], [14, 13], [15, 16], [16, 11], [16, 15],
+                [16, 17], [17, 16], [17, 18], [18, 13], [18, 17], [18, 19],
                 [19, 18]]
 
         configuration = QasmBackendConfiguration(
-            backend_name='fake_poughkeepsie',
+            backend_name='fake_boeblingen',
             backend_version='0.0.0',
             n_qubits=20,
             basis_gates=['u1', 'u2', 'u3', 'cx', 'id'],
@@ -65,7 +67,7 @@ class FakeJohannesburg(FakeBackend):
         """Returns a snapshot of device properties as recorded on 10/08/19.
         """
         dirname = os.path.dirname(__file__)
-        filename = "props_poughkeepsie.json"
+        filename = "props_boeblingen.json"
         with open(os.path.join(dirname, filename), "r") as f_prop:
             props = json.load(f_prop)
         return BackendProperties.from_dict(props)
