+++ qiskit/test/mock/backends/boeblingen/props_boeblingen.json
