+++ qiskit/test/mock/backends/burlington/props_burlington.json
