+++ qiskit/test/mock/backends/johannesburg/props_johannesburg.json
