+++ qiskit/test/mock/backends/melbourne/fake_melbourne.py
@@ -13,7 +13,7 @@
 # that they have been altered from the originals.
 
 """
-Fake Boeblingen device (20 qubit).
+Fake Melbourne device (14 qubit).
 """
 
 import os
@@ -21,42 +21,33 @@ import json
 
 from qiskit.providers.models import (GateConfig, QasmBackendConfiguration,
                                      BackendProperties)
-from .fake_backend import FakeBackend
+from qiskit.test.mock.fake_backend import FakeBackend
 
 
-class FakeBoeblingen(FakeBackend):
-    """A fake Boeblingen backend."""
+class FakeMelbourne(FakeBackend):
+    """A fake 14 qubit backend."""
 
     def __init__(self):
         """
-          00 ↔ 01 ↔ 02 ↔ 03 ↔ 04
-                ↕         ↕
-          05 ↔ 06 ↔ 07 ↔ 08 ↔ 09
-           ↕         ↕         ↕
-          10 ↔ 11 ↔ 12 ↔ 13 ↔ 14
-                ↕         ↕
-          15 ↔ 16 ↔ 17 ↔ 18 ↔ 19
+         0 ← 1 →  2 →  3 ←  4 ← 5 → 6
+             ↑    ↑    ↑    ↓   ↓   ↓
+            13 → 12 ← 11 → 10 ← 9 → 8 ← 7
         """
-        cmap = [[0, 1], [1, 0], [1, 2], [1, 6], [2, 1], [2, 3], [3, 2],
-                [3, 4], [3, 8], [4, 3], [5, 6], [5, 10], [6, 1], [6, 5],
-                [6, 7], [7, 6], [7, 8], [7, 12], [8, 3], [8, 7], [8, 9],
-                [9, 8], [9, 14], [10, 5], [10, 11], [11, 10], [11, 12],
-                [11, 16], [12, 7], [12, 11], [12, 13], [13, 12], [13, 14],
-                [13, 18], [14, 9], [14, 13], [15, 16], [16, 11], [16, 15],
-                [16, 17], [17, 16], [17, 18], [18, 13], [18, 17], [18, 19],
-                [19, 18]]
+        cmap = [[1, 0], [1, 2], [2, 3], [4, 3], [4, 10], [5, 4],
+                [5, 6], [5, 9], [6, 8], [7, 8], [9, 8], [9, 10],
+                [11, 3], [11, 10], [11, 12], [12, 2], [13, 1], [13, 12]]
 
         configuration = QasmBackendConfiguration(
-            backend_name='fake_boeblingen',
+            backend_name='fake_melbourne',
             backend_version='0.0.0',
-            n_qubits=20,
+            n_qubits=14,
             basis_gates=['u1', 'u2', 'u3', 'cx', 'id'],
             simulator=False,
             local=True,
             conditional=False,
             open_pulse=False,
-            memory=True,
-            max_shots=8192,
+            memory=False,
+            max_shots=65536,
             gates=[GateConfig(name='TODO', parameters=[], qasm_def='TODO')],
             coupling_map=cmap,
         )
@@ -64,10 +55,10 @@ class FakeBoeblingen(FakeBackend):
         super().__init__(configuration)
 
     def properties(self):
-        """Returns a snapshot of device properties as recorded on 10/08/19.
+        """Returns a snapshot of device properties as recorded on 8/30/19.
         """
         dirname = os.path.dirname(__file__)
-        filename = "props_boeblingen.json"
+        filename = "props_melbourne.json"
         with open(os.path.join(dirname, filename), "r") as f_prop:
             props = json.load(f_prop)
         return BackendProperties.from_dict(props)
