+++ qiskit/test/mock/backends/ourense/props_ourense.json
