+++ qiskit/test/mock/backends/poughkeepsie/fake_poughkeepsie.py
@@ -13,7 +13,7 @@
 # that they have been altered from the originals.
 
 """
-Fake Melbourne device (14 qubit).
+Fake Poughkeepsie device (20 qubit).
 """
 
 import os
@@ -21,33 +21,40 @@ import json
 
 from qiskit.providers.models import (GateConfig, QasmBackendConfiguration,
                                      BackendProperties)
-from .fake_backend import FakeBackend
+from qiskit.test.mock.fake_backend import FakeBackend
 
 
-class FakeMelbourne(FakeBackend):
-    """A fake 14 qubit backend."""
+class FakePoughkeepsie(FakeBackend):
+    """A fake Poughkeepsie backend."""
 
     def __init__(self):
         """
-         0 ← 1 →  2 →  3 ←  4 ← 5 → 6
-             ↑    ↑    ↑    ↓   ↓   ↓
-            13 → 12 ← 11 → 10 ← 9 → 8 ← 7
+          00 ↔ 01 ↔ 02 ↔ 03 ↔ 04
+           ↕                   ↕
+          05 ↔ 06 ↔ 07 ↔ 08 ↔ 09
+           ↕         ↕         ↕
+          10 ↔ 11 ↔ 12 ↔ 13 ↔ 14
+           ↕                   ↕
+          15 ↔ 16 ↔ 17 ↔ 18 ↔ 19
         """
-        cmap = [[1, 0], [1, 2], [2, 3], [4, 3], [4, 10], [5, 4],
-                [5, 6], [5, 9], [6, 8], [7, 8], [9, 8], [9, 10],
-                [11, 3], [11, 10], [11, 12], [12, 2], [13, 1], [13, 12]]
+        cmap = [[0, 1], [0, 5], [1, 0], [1, 2], [2, 1], [2, 3], [3, 2], [3, 4], [4, 3], [4, 9],
+                [5, 0], [5, 6], [5, 10], [6, 5], [6, 7], [7, 6], [7, 8], [7, 12], [8, 7], [8, 9],
+                [9, 4], [9, 8], [9, 14], [10, 5], [10, 11], [10, 15], [11, 10], [11, 12], [12, 7],
+                [12, 11], [12, 13], [13, 12], [13, 14], [14, 9], [14, 13], [14, 19], [15, 10],
+                [15, 16], [16, 15], [16, 17], [17, 16], [17, 18], [18, 17], [18, 19], [19, 14],
+                [19, 18]]
 
         configuration = QasmBackendConfiguration(
-            backend_name='fake_melbourne',
+            backend_name='fake_poughkeepsie',
             backend_version='0.0.0',
-            n_qubits=14,
+            n_qubits=20,
             basis_gates=['u1', 'u2', 'u3', 'cx', 'id'],
             simulator=False,
             local=True,
             conditional=False,
             open_pulse=False,
-            memory=False,
-            max_shots=65536,
+            memory=True,
+            max_shots=8192,
             gates=[GateConfig(name='TODO', parameters=[], qasm_def='TODO')],
             coupling_map=cmap,
         )
@@ -55,10 +62,10 @@ class FakeMelbourne(FakeBackend):
         super().__init__(configuration)
 
     def properties(self):
-        """Returns a snapshot of device properties as recorded on 8/30/19.
+        """Returns a snapshot of device properties as recorded on 10/08/19.
         """
         dirname = os.path.dirname(__file__)
-        filename = "props_melbourne.json"
+        filename = "props_poughkeepsie.json"
         with open(os.path.join(dirname, filename), "r") as f_prop:
             props = json.load(f_prop)
         return BackendProperties.from_dict(props)
