+++ qiskit/test/mock/backends/poughkeepsie/props_poughkeepsie.json
