+++ qiskit/test/mock/backends/singapore/fake_singapore.py
@@ -13,7 +13,7 @@
 # that they have been altered from the originals.
 
 """
-Fake Tenerife device (5 qubit).
+Fake Boeblingen device (20 qubit).
 """
 
 import os
@@ -21,33 +21,42 @@ import json
 
 from qiskit.providers.models import (GateConfig, QasmBackendConfiguration,
                                      BackendProperties)
-from .fake_backend import FakeBackend
+from qiskit.test.mock.fake_backend import FakeBackend
 
 
-class FakeTenerife(FakeBackend):
-    """A fake 5 qubit backend."""
+class FakeSingapore(FakeBackend):
+    """A fake Singapore backend."""
 
     def __init__(self):
         """
-             1
-           ↙ ↑
-         0 ← 2 ← 3
-             ↑ ↙
-             4
+          00 ↔ 01 ↔ 02 ↔ 03 ↔ 04
+                ↕         ↕
+          05 ↔ 06 ↔ 07 ↔ 08 ↔ 09
+           ↕         ↕         ↕
+          10 ↔ 11 ↔ 12 ↔ 13 ↔ 14
+                ↕         ↕
+          15 ↔ 16 ↔ 17 ↔ 18 ↔ 19
         """
-        cmap = [[1, 0], [2, 0], [2, 1], [3, 2], [3, 4], [4, 2]]
+        cmap = [[0, 1], [1, 0], [1, 2], [1, 6], [2, 1], [2, 3], [3, 2],
+                [3, 4], [3, 8], [4, 3], [5, 6], [5, 10], [6, 1], [6, 5],
+                [6, 7], [7, 6], [7, 8], [7, 12], [8, 3], [8, 7], [8, 9],
+                [9, 8], [9, 14], [10, 5], [10, 11], [11, 10], [11, 12],
+                [11, 16], [12, 7], [12, 11], [12, 13], [13, 12], [13, 14],
+                [13, 18], [14, 9], [14, 13], [15, 16], [16, 11], [16, 15],
+                [16, 17], [17, 16], [17, 18], [18, 13], [18, 17], [18, 19],
+                [19, 18]]
 
         configuration = QasmBackendConfiguration(
-            backend_name='fake_tenerife',
+            backend_name='fake_singapore',
             backend_version='0.0.0',
-            n_qubits=5,
+            n_qubits=20,
             basis_gates=['u1', 'u2', 'u3', 'cx', 'id'],
             simulator=False,
             local=True,
             conditional=False,
             open_pulse=False,
-            memory=False,
-            max_shots=65536,
+            memory=True,
+            max_shots=8192,
             gates=[GateConfig(name='TODO', parameters=[], qasm_def='TODO')],
             coupling_map=cmap,
         )
@@ -55,10 +64,10 @@ class FakeTenerife(FakeBackend):
         super().__init__(configuration)
 
     def properties(self):
-        """Returns a snapshot of device properties as recorded on 8/30/19.
+        """Returns a snapshot of device properties as recorded on 10/08/19.
         """
         dirname = os.path.dirname(__file__)
-        filename = "props_tenerife.json"
+        filename = "props_singapore.json"
         with open(os.path.join(dirname, filename), "r") as f_prop:
             props = json.load(f_prop)
         return BackendProperties.from_dict(props)
