+++ qiskit/test/mock/backends/singapore/props_singapore.json
