+++ qiskit/test/mock/backends/tenerife/props_tenerife.json
