+++ qiskit/test/mock/backends/tokyo/props_tokyo.json
