+++ qiskit/test/mock/backends/vigo/props_vigo.json
