+++ qiskit/test/mock/backends/yorktown/props_yorktown.json
