+++ qiskit/test/mock/fake_provider.py
@@ -19,17 +19,7 @@ Fake provider class that provides access to fake backends.
 from qiskit.providers.baseprovider import BaseProvider
 from qiskit.providers.exceptions import QiskitBackendNotFoundError
 
-from .fake_qasm_simulator import FakeQasmSimulator
-from .fake_yorktown import FakeYorktown
-from .fake_tenerife import FakeTenerife
-from .fake_ourense import FakeOurense
-from .fake_melbourne import FakeMelbourne
-from .fake_rueschlikon import FakeRueschlikon
-from .fake_tokyo import FakeTokyo
-from .fake_poughkeepsie import FakePoughkeepsie
-from .fake_boeblingen import FakeBoeblingen
-from .fake_openpulse_2q import FakeOpenPulse2Q
-from .fake_openpulse_3q import FakeOpenPulse3Q
+from .backends import *
 
 
 class FakeProvider(BaseProvider):
@@ -60,10 +50,15 @@ class FakeProvider(BaseProvider):
                           FakeYorktown(),
                           FakeTenerife(),
                           FakeOurense(),
+                          FakeVigo(),
+                          FakeBurlington(),
                           FakeMelbourne(),
                           FakeRueschlikon(),
                           FakeTokyo(),
                           FakePoughkeepsie(),
+                          FakeAlmaden(),
+                          FakeSingapore(),
+                          FakeJohannesburg(),
                           FakeBoeblingen()]
 
         super().__init__()
